+++ R/natural_join.R
@@ -61,6 +61,7 @@ natural_join <- function(a, b,
                          by,
                          jointype = 'INNER',
                          env = parent.frame()) {
+  force(env)
   UseMethod("natural_join", a)
 }
 
@@ -70,6 +71,7 @@ natural_join.relop <- function(a, b,
                                by,
                                jointype = 'INNER',
                                env = parent.frame()) {
+  force(env)
   if(length(list(...))>0) {
     stop("rquery::natural_join unexpected arguments")
   }
@@ -101,6 +103,7 @@ natural_join.data.frame <- function(a, b,
                                     by,
                                     jointype = 'INNER',
                                     env = parent.frame()) {
+  force(env)
   if(length(list(...))>0) {
     stop("rquery::natural_join unexpected arguments")
   }
@@ -114,7 +117,8 @@ natural_join.data.frame <- function(a, b,
   dnodeb <- mk_td(tmp_namea, colnames(b))
   enode <- natural_join(dnodea, dnodeb,
                         jointype = jointype,
-                        by = by)
+                        by = by,
+                        env = env)
   source <- list(a = a, b = b)
   names(source) <- c(tmp_namea, tmp_nameb)
   rquery_apply_to_data_frame(source, enode, env = env)