+++ R/rename.R
@@ -24,12 +24,14 @@
 #'
 rename_columns <- function(source, cmap,
                            env = parent.frame()) {
+  force(env)
   UseMethod("rename_columns", source)
 }
 
 #' @export
 rename_columns.relop <- function(source, cmap,
                                  env = parent.frame()) {
+  force(env)
   if(length(cmap)<=0) {
     stop("rquery::rename_columns must rename at least 1 column")
   }
@@ -59,6 +61,7 @@ rename_columns.relop <- function(source, cmap,
 #' @export
 rename_columns.data.frame <- function(source, cmap,
                                       env = parent.frame()) {
+  force(env)
   if(length(cmap)<=0) {
     stop("rquery::rename_columns must rename at least 1 column")
   }
@@ -70,7 +73,7 @@ rename_columns.data.frame <- function(source, cmap,
   }
   tmp_name <- mk_tmp_name_source("rquery_tmp")()
   dnode <- mk_td(tmp_name, colnames(source))
-  enode <- rename_columns(dnode, cmap)
+  enode <- rename_columns(dnode, cmap, env = env)
   rquery_apply_to_data_frame(source, enode, env = env)
 }