+++ NEWS.md
@@ -1,4 +1,10 @@
 
+# rquery 1.1.0 2018/09/20
+
+ * Altrenate data.table implementation path.
+ * lookup_by_column().
+ * Force parent.frame().
+
 # rquery 1.0.0 2018/09/10
 
  * Fix key_inspector_postgresql quoting issue.